+++ src/helixnet/models.py
@@ -4,7 +4,7 @@ from typing import List, Tuple, Dict
 
 
 import mygrad as mg
-
+import numpy as np
 
 from helixnet import layers
 
@@ -13,10 +13,6 @@ class Sequential:
     """A Simple model that propagate through the layers in a linear way
     
     :param list layer: the list which contains the layers"""
-
-    def __init__(self, layers_: list[layers.Layer]) -> None:
-        self.layers = layers_
-
 
     def __init__(self, layers_: list[layers.Layer]) -> None:
         self.layers = layers_
@@ -66,7 +62,6 @@ class Sequential:
         self.layers.append(layer)
 
     def summary(self) -> None:
-
         """
         This method prints the model summary which contains
         the name of every layer and it's shape
@@ -93,6 +88,20 @@ class Sequential:
             x = layer.predict(x)
         return x
 
-    def set_weights(self, weights: list[mg.Tensor]):
-        for layer, weight in zip(self.layers, weights):
-            layer.set_weights(weight)
+    def get_weights(self) -> List[np.ndarray]:
+        """Returns a flat list of all trainable weights in the model."""
+        return [weight for layer in self.layers for weight in layer.get_weights()]
+
+    def set_weights(self, weights: List[np.ndarray]):
+        """
+        Sets the model's weights from a flat list.
+
+        :param List[np.ndarray] weights: The weights what will be produced by
+            :func:`helixnet.io.load_model`
+        """
+        weight_iterator = iter(weights)
+        for layer in self.layers:
+            # Slicing the iterator to get the correct number of weights for this layer
+            num_params = len(layer.trainable_params)
+            layer_weights = [next(weight_iterator) for _ in range(num_params)]
+            layer.set_weights(layer_weights)