+++ tests/__init__.py
@@ -1,2 +1,3 @@
 from . import layers_test
-from . import models_test
+from . import io_test
+from . import optimizers_test